requirements.txt
requests
trueskill
django
python-dotenv